+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Implement connection throttling (i: @mitchhentges, r: @nerdbeere)
+
 - Fix a bug that prevented the user from adding mappings (i: @nerdbeere, r: @davidneat)
 - Add mapping active (i: @davidneat, r: @nerdbeere)
 - Add remove mapping to context menu (i: @davidneat, r: @nerdbeere)